app/requirements.txt
openai>=1.30.0
faiss-cpu>=1.7.4
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
tiktoken>=0.6.0
sentence-transformers>=3.0.0
